Updated audio and shop backend integration
